Server recieves and transmits file data as needed and provides atk param
The server now is able to take the file from the client and store it,
when a node checks with the server it issues the node an attack ID, a
sequence to test and the file data itself which is then reconstructed on
the node to be attacked. Further testing of success scenarios,
cancellation, node fall over and secondary upload of target need to be
handles and tested.