Health check of nodes, success receive and transmission
The health check system has been configured and enabled. This now logs
all the attached nodes and tracks if they are alive, their individual ID
and the attack sequence assigned to them. If they are marked as DEAD due
to a period without PING, they are removed from the server as a node and
their sequence put up for retrieval by the next available node (even if
it is the same node reconnecting). 

In the event a node retrieves the correct password the attack is
terminated and the nodes are informed. The client then can then retrieve
the password by polling the server.