Version 1.5
System is now fully integrated with other components.Attack methods both
brute and dictionary are available with handling for each method
relevant to their use. Tested with both the LNX and Windows nodes with
Windows client and methods are successful with attacks working as
expected. 

Only flaw is when dictionary attack is running and node downs, other
nodes will not try to pick up the slack unless they are also running.
This may need addressing.