Version 2.0
Balance and dictionary attack tested with health checks and safe guards
in place. System has been fully tested. 

Additional work - Create a fail safe feature for the server to recover
after crash by checking related database for last attack information.
This will require current ARN, balance information, type of attack and a
check as to whether the file in question is still in the location
specified (fragment is still in directory).